separates participant popover
